moved db folder to src
