+++ src/pages/inventory/index.tsx
@@ -1,8 +1,7 @@
 import Layout from "@/components/layout/Layout";
 import InventoryProduct from "@/components/inventory";
-
 import useSWR from "swr";
-import { ProductType } from "../../../db/model/Product";
+import { ProductType } from "@/db/model/Product";
 import PDFGenerator from "@/components/pdf-generator";
 
 export default function InventoryPage() {
@@ -19,7 +18,7 @@ export default function InventoryPage() {
     <>
       <h1>Inventory</h1>
       <div>
-        {products.map((product: ProductType) => {
+        {products?.map((product: ProductType) => {
           return (
             <div key={product._id}>
               <InventoryProduct product={product} mutate={mutate} />
